Dynamic version of create_warp_tensors.
This patch removes the static sized array by dynamically constructing a
chain of lambdas that resolves to WMMA structs when tile_x and tile_y
are statically known. Think of this as implementing a linked list
through lambda expressions.
For some reason, it is not possible to direclty construct this as a two
dimensional grid in two nested unrolls; thorin creates closures if that
is attempted. Consequently, we stick to a one dimensional list and map
it to "x,y"-space in a separate step.